registry code format

+++ mvm/quorum/contracts/registry/contracts/Registry.sol
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.8.0 <0.9.0;
 
-import {Bytes} from './libs/Bytes.sol';
-import {BLS} from './libs/BLS.sol';
-import {Storage} from './Storage.sol';
-import {IRegistry,Registrable} from './Registrable.sol';
-import {Asset} from './Asset.sol';
-import {User} from './User.sol';
+import {Bytes} from "./libs/Bytes.sol";
+import {BLS} from "./libs/BLS.sol";
+import {Storage} from "./Storage.sol";
+import {IRegistry, Registrable} from "./Registrable.sol";
+import {Asset} from "./Asset.sol";
+import {User} from "./User.sol";
 
 contract Registry is IRegistry {
     using Bytes for bytes;
@@ -14,7 +14,7 @@ contract Registry is IRegistry {
     using BLS for bytes;
 
     event UserCreated(address at, bytes members);
-    event AssetCreated(address at, uint id);
+    event AssetCreated(address at, uint256 id);
     event Halted(bool state);
     event Iterated(uint256[4] from, uint256[4] to);
     event MixinTransaction(bytes);
@@ -31,7 +31,7 @@ contract Registry is IRegistry {
 
     mapping(address => bytes) public users;
     mapping(address => uint128) public assets;
-    mapping(uint => address) public contracts;
+    mapping(uint256 => address) public contracts;
     mapping(uint128 => uint256) public balances;
     address[] public addresses;
     uint128[] public deposits;
@@ -61,7 +61,12 @@ contract Registry is IRegistry {
     function iterate(bytes memory raw) public {
         require(HALTED, "invalid state");
         require(raw.length == 256, "invalid input size");
-        uint256[4] memory group = [raw.toUint256(0), raw.toUint256(32), raw.toUint256(64), raw.toUint256(96)];
+        uint256[4] memory group = [
+            raw.toUint256(0),
+            raw.toUint256(32),
+            raw.toUint256(64),
+            raw.toUint256(96)
+        ];
         uint256[2] memory sig1 = [raw.toUint256(128), raw.toUint256(160)];
         uint256[2] memory sig2 = [raw.toUint256(192), raw.toUint256(224)];
         uint256[2] memory message = raw.slice(0, 128).hashToPoint();
@@ -98,17 +103,33 @@ contract Registry is IRegistry {
         return true;
     }
 
-    function sendMixinTransaction(address user, address asset, uint256 amount) internal {
+    function sendMixinTransaction(
+        address user,
+        address asset,
+        uint256 amount
+    ) internal {
         uint256 balance = balances[assets[asset]];
         bytes memory extra = new bytes(0);
-        bytes memory log = buildMixinTransaction(OUTBOUND, users[user], assets[asset], amount, extra);
+        bytes memory log = buildMixinTransaction(
+            OUTBOUND,
+            users[user],
+            assets[asset],
+            amount,
+            extra
+        );
         emit MixinTransaction(log);
         balances[assets[asset]] = balance - amount;
         OUTBOUND = OUTBOUND + 1;
     }
 
     // process || nonce || asset || amount || extra || timestamp || members || threshold || sig
-    function buildMixinTransaction(uint64 nonce, bytes memory receiver, uint128 asset, uint256 amount, bytes memory extra) internal view returns (bytes memory) {
+    function buildMixinTransaction(
+        uint64 nonce,
+        bytes memory receiver,
+        uint128 asset,
+        uint256 amount,
+        bytes memory extra
+    ) internal view returns (bytes memory) {
         require(extra.length < 128, "extra too large");
         bytes memory raw = uint128ToFixedBytes(PID);
         raw = raw.concat(uint64ToFixedBytes(nonce));
@@ -147,8 +168,11 @@ contract Registry is IRegistry {
         (evt.asset, evt.extra) = parseEventInput(id, evt.extra);
 
         offset = offset + 2;
-        evt.sig = [raw.toUint256(offset), raw.toUint256(offset+32)];
-        uint256[2] memory message = raw.slice(0, offset-2).concat(new bytes(2)).hashToPoint();
+        evt.sig = [raw.toUint256(offset), raw.toUint256(offset + 32)];
+        uint256[2] memory message = raw
+            .slice(0, offset - 2)
+            .concat(new bytes(2))
+            .hashToPoint();
         require(evt.sig.verifySingle(GROUP, message), "invalid signature");
 
         offset = offset + 64;
@@ -166,8 +190,16 @@ contract Registry is IRegistry {
         return User(evt.user).run(evt.asset, evt.amount, evt.extra);
     }
 
-    function parseEventExtra(bytes memory raw, uint offset) internal pure returns(uint, bytes memory, uint64) {
-        uint size = raw.toUint16(offset);
+    function parseEventExtra(bytes memory raw, uint256 offset)
+        internal
+        pure
+        returns (
+            uint256,
+            bytes memory,
+            uint64
+        )
+    {
+        uint256 size = raw.toUint16(offset);
         offset = offset + 2;
         bytes memory extra = raw.slice(offset, size);
         offset = offset + size;
@@ -176,19 +208,32 @@ contract Registry is IRegistry {
         return (offset, extra, timestamp);
     }
 
-    function parseEventAsset(bytes memory raw, uint offset) internal pure returns(uint, uint128, uint256) {
+    function parseEventAsset(bytes memory raw, uint256 offset)
+        internal
+        pure
+        returns (
+            uint256,
+            uint128,
+            uint256
+        )
+    {
         uint128 id = raw.toUint128(offset);
         require(id > 0, "invalid asset");
         offset = offset + 16;
-        uint size = raw.toUint16(offset);
+        uint256 size = raw.toUint16(offset);
         offset = offset + 2;
         require(size <= 32, "integer out of bounds");
-        uint256 amount = new bytes(32 - size).concat(raw.slice(offset, size)).toUint256(0);
+        uint256 amount = new bytes(32 - size)
+            .concat(raw.slice(offset, size))
+            .toUint256(0);
         offset = offset + size;
         return (offset, id, amount);
     }
 
-    function parseEventUser(bytes memory raw, uint offset) internal returns (uint, address) {
+    function parseEventUser(bytes memory raw, uint256 offset)
+        internal
+        returns (uint256, address)
+    {
         uint16 size = raw.toUint16(offset);
         size = 2 + size * 16 + 2;
         bytes memory members = raw.slice(offset, size);
@@ -197,8 +242,11 @@ contract Registry is IRegistry {
         return (offset, user);
     }
 
-    function parseEventInput(uint128 id, bytes memory extra) internal returns (address, bytes memory) {
-        uint offset = 0;
+    function parseEventInput(uint128 id, bytes memory extra)
+        internal
+        returns (address, bytes memory)
+    {
+        uint256 offset = 0;
         uint16 size = extra.toUint16(offset);
         offset = offset + 2;
         string memory symbol = string(extra.slice(offset, size));
@@ -215,7 +263,11 @@ contract Registry is IRegistry {
         return (asset, input);
     }
 
-    function getOrCreateAssetContract(uint128 id, string memory symbol, string memory name) internal returns (address) {
+    function getOrCreateAssetContract(
+        uint128 id,
+        string memory symbol,
+        string memory name
+    ) internal returns (address) {
         address old = contracts[id];
         if (old != address(0)) {
             return old;
@@ -234,8 +286,11 @@ contract Registry is IRegistry {
         return asset;
     }
 
-    function getOrCreateUserContract(bytes memory members) internal returns (address) {
-        uint id = uint256(keccak256(members));
+    function getOrCreateUserContract(bytes memory members)
+        internal
+        returns (address)
+    {
+        uint256 id = uint256(keccak256(members));
         address old = contracts[id];
         if (old != address(0)) {
             return old;
@@ -254,24 +309,44 @@ contract Registry is IRegistry {
         return user;
     }
 
-    function getUserContractCode(bytes memory members) internal pure returns (bytes memory) {
+    function getUserContractCode(bytes memory members)
+        internal
+        pure
+        returns (bytes memory)
+    {
         bytes memory code = type(User).creationCode;
         bytes memory args = abi.encode(members);
         return abi.encodePacked(code, args);
     }
 
-    function getAssetContractCode(uint id, string memory symbol, string memory name) internal pure returns (bytes memory) {
+    function getAssetContractCode(
+        uint256 id,
+        string memory symbol,
+        string memory name
+    ) internal pure returns (bytes memory) {
         bytes memory code = type(Asset).creationCode;
         bytes memory args = abi.encode(id, name, symbol);
         return abi.encodePacked(code, args);
     }
 
-    function getContractAddress(bytes memory code) internal view returns (address) {
-        code = abi.encodePacked(bytes1(0xff), address(this), VERSION, keccak256(code));
-        return address(uint160(uint(keccak256(code))));
+    function getContractAddress(bytes memory code)
+        internal
+        view
+        returns (address)
+    {
+        code = abi.encodePacked(
+            bytes1(0xff),
+            address(this),
+            VERSION,
+            keccak256(code)
+        );
+        return address(uint160(uint256(keccak256(code))));
     }
 
-    function deploy(bytes memory bytecode, uint _salt) internal returns (address) {
+    function deploy(bytes memory bytecode, uint256 _salt)
+        internal
+        returns (address)
+    {
         address addr;
         assembly {
             addr := create2(
@@ -288,11 +363,10 @@ contract Registry is IRegistry {
         return addr;
     }
 
-
     function uint16ToFixedBytes(uint16 x) internal pure returns (bytes memory) {
         bytes memory c = new bytes(2);
         bytes2 b = bytes2(x);
-        for (uint i=0; i < 2; i++) {
+        for (uint256 i = 0; i < 2; i++) {
             c[i] = b[i];
         }
         return c;
@@ -301,32 +375,40 @@ contract Registry is IRegistry {
     function uint64ToFixedBytes(uint64 x) internal pure returns (bytes memory) {
         bytes memory c = new bytes(8);
         bytes8 b = bytes8(x);
-        for (uint i=0; i < 8; i++) {
+        for (uint256 i = 0; i < 8; i++) {
             c[i] = b[i];
         }
         return c;
     }
 
-    function uint128ToFixedBytes(uint128 x) internal pure returns (bytes memory) {
+    function uint128ToFixedBytes(uint128 x)
+        internal
+        pure
+        returns (bytes memory)
+    {
         bytes memory c = new bytes(16);
         bytes16 b = bytes16(x);
-        for (uint i=0; i < 16; i++) {
+        for (uint256 i = 0; i < 16; i++) {
             c[i] = b[i];
         }
         return c;
     }
 
-    function uint256ToVarBytes(uint256 x) internal pure returns (bytes memory, uint16) {
+    function uint256ToVarBytes(uint256 x)
+        internal
+        pure
+        returns (bytes memory, uint16)
+    {
         bytes memory c = new bytes(32);
         bytes32 b = bytes32(x);
         uint16 offset = 0;
-        for (uint16 i=0; i < 32; i++) {
+        for (uint16 i = 0; i < 32; i++) {
             c[i] = b[i];
             if (c[i] > 0 && offset == 0) {
                 offset = i;
             }
         }
         uint16 size = 32 - offset;
-        return (c.slice(offset, 32-offset), size);
+        return (c.slice(offset, 32 - offset), size);
     }
 }
